+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - chore: rename agent lifecycle methods and APIs [#858](https://github.com/hypermodeinc/modus/pull/858)
 - feat: enforce WASI reactor mode [#859](https://github.com/hypermodeinc/modus/pull/859)
 - feat: return user and chat errors in API response [#863](https://github.com/hypermodeinc/modus/pull/863)
+- feat: list agents on health endpoint [#865](https://github.com/hypermodeinc/modus/pull/865)
 
 ## 2025-05-22 - Go SDK 0.18.0-alpha.3